Simplify and move all commands to components
