+++ autogpts/autogpt/agbenchmark_config/benchmarks.py
@@ -6,11 +6,9 @@ from pathlib import Path
 from autogpt.agent_manager.agent_manager import AgentManager
 from autogpt.agents.agent import Agent, AgentConfiguration, AgentSettings
 from autogpt.app.main import _configure_openai_provider, run_interaction_loop
-from autogpt.commands import COMMAND_CATEGORIES
 from autogpt.config import AIProfile, ConfigBuilder
 from autogpt.file_storage import FileStorageBackendName, get_storage
 from autogpt.logs.config import configure_logging
-from autogpt.models.command_registry import CommandRegistry
 
 LOG_DIR = Path(__file__).parent / "logs"
 
@@ -31,8 +29,6 @@ def bootstrap_agent(task: str, continuous_mode: bool) -> Agent:
     config.continuous_limit = 20
     config.noninteractive_mode = True
     config.memory_backend = "no_memory"
-
-    command_registry = CommandRegistry.with_command_modules(COMMAND_CATEGORIES, config)
 
     ai_profile = AIProfile(
         ai_name="AutoGPT",
@@ -68,7 +64,6 @@ def bootstrap_agent(task: str, continuous_mode: bool) -> Agent:
     agent = Agent(
         settings=agent_settings,
         llm_provider=_configure_openai_provider(config),
-        command_registry=command_registry,
         file_storage=file_storage,
         legacy_config=config,
     )
