+++ autogpts/autogpt/autogpt/agent_factory/configurators.py
@@ -1,12 +1,10 @@
 from typing import Optional
 
 from autogpt.agents.agent import Agent, AgentConfiguration, AgentSettings
-from autogpt.commands import COMMAND_CATEGORIES
 from autogpt.config import AIDirectives, AIProfile, Config
 from autogpt.core.resource.model_providers import ChatModelProvider
 from autogpt.file_storage.base import FileStorage
 from autogpt.logs.config import configure_chat_plugins
-from autogpt.models.command_registry import CommandRegistry
 from autogpt.plugins import scan_plugins
