--- autogpts/autogpt/autogpt/commands/__init__.py
@@ -1,7 +1,0 @@
-COMMAND_CATEGORIES = [
-    "autogpt.commands.execute_code",
-    "autogpt.commands.file_operations",
-    "autogpt.commands.web_search",
-    "autogpt.commands.web_selenium",
-    "autogpt.commands.image_gen",
-]
