+++ autogpts/autogpt/tests/integration/test_image_gen.py
@@ -7,7 +7,7 @@ import pytest
 from PIL import Image
 
 from autogpt.agents.agent import Agent
-from autogpt.commands.image_gen import generate_image, generate_image_with_sd_webui
+from autogpt.components.image_gen import generate_image, generate_image_with_sd_webui
 
 
 @pytest.fixture(params=[256, 512, 1024])
