+++ autogpts/autogpt/tests/unit/test_git_commands.py
@@ -4,7 +4,7 @@ from git.repo.base import Repo
 
 from autogpt.agents.agent import Agent
 from autogpt.agents.utils.exceptions import CommandExecutionError
-from autogpt.commands.git_operations import clone_repository
+from autogpt.components.git_operations import clone_repository
 from autogpt.file_storage.base import FileStorage
