+++ autogpts/autogpt/tests/unit/test_web_search.py
@@ -5,7 +5,7 @@ from googleapiclient.errors import HttpError
 
 from autogpt.agents.agent import Agent
 from autogpt.agents.utils.exceptions import ConfigurationError
-from autogpt.commands.web_search import google, safe_google_results, web_search
+from autogpt.components.web_search import google, safe_google_results, web_search
 
 
 @pytest.mark.parametrize(
